feat(person): refactory scripts folder
