+++ src/pages/person/person-list/person-list.js
@@ -1,4 +1,4 @@
-import { PersonController } from '../person-controller.js'
+import { PersonController } from '../scripts/person-controller.js'
 
 let controller = new PersonController()
 controller.bindTable()
